+++ test_assets/features/step_definitions/passing.js
@@ -17,6 +17,19 @@ module.exports = function() {
     callback('Failed');
   });
 
+  this.When(/^an action is executed that passes on retry '(.*)'$/, function(retryCount, callback) {
+    if (this.isDryRun()) { return callback(); }
+
+    var currentRetryCount = parseInt(process.env.PARALLEL_CUCUMBER_RETRY);
+
+    if (currentRetryCount < retryCount) {
+      callback('Failed on retry ' + currentRetryCount);
+    }
+    else {
+      callback();
+    }
+  });
+
   this.Then(/^a post-condition passes$/, function(callback) {
     if (this.isDryRun()) { return callback(); }